requirements.txt
flask==2.3.3
flask-sqlalchemy==3.1.1
flask-migrate==4.0.5
flask-wtf==1.2.1
flask-login==0.6.3
flask-cors==4.0.0
celery==5.3.4
redis==5.0.1
boto3==1.34.1
python-dotenv==1.0.0
gunicorn==21.2.0
pillow==10.1.0
ffmpeg-python==0.2.0
psutil==5.9.6
werkzeug==2.3.7
pydantic==2.5.2
marshmallow==3.20.1
python-magic==0.4.27
watchdog==3.0.0
PyJWT==2.8.0
flower==2.0.1